+++ lib/trader.py
@@ -1,5 +1,5 @@
-from src.experience_replay import HERBuffer
-from src.state import State
+from lib.experience_replay import HERBuffer
+from lib.state import State
 import numpy as np
 import tensorflow as tf
 import random
